requirements.txt
aiobotocore==2.13.0
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
alabaster==0.7.16
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
babel==2.15.0
backoff==2.2.1
backports-tarfile==1.1.1
beautifulsoup4==4.12.3
bidsschematools==0.7.2
bleach==6.1.0
botocore==1.34.106
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click-didyoumean==0.3.1
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
dandi==0.62.1
dandischema==0.10.1
defusedxml==0.7.1
dnspython==2.6.1
docutils==0.21.2
email-validator==2.1.1
etelemetry==0.3.1
exceptiongroup==1.2.1
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.16.0
fonttools==4.52.1
fqdn==1.5.1
frozenlist==1.4.1
fscacher==0.4.0
fsspec==2024.5.0
gast==0.4.0
greenlet==3.1.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.7
imageio==2.34.1
imagesize==1.4.1
importlib-metadata==4.13.0
importlib-resources==6.4.0
iniconfig==2.0.0
interleave==0.2.1
ipydatawidgets==4.3.2
ipyfilechooser==0.6.0
ipympl==0.9.4
ipython==8.18.1
ipython-genutils==0.2.0
ipyvolume==0.6.3
ipyvue==1.11.1
ipyvuetify==1.9.4
ipywebrtc==0.6.0
ipywidgets==8.1.2
isodate==0.6.1
isoduration==20.11.0
jaraco-functools==4.0.1
jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.1
jupyter-server-terminals==0.5.3
jupyterlab==4.2.1
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.2
jupyterlab-widgets==3.0.10
keyring==25.2.1
keyrings-alt==5.0.1
kiwisolver==1.4.5
lazy-loader==0.4
llvmlite==0.42.0
mako==1.3.5
markupsafe==2.1.5
matplotlib==3.9.0
mistune==3.0.2
more-itertools==10.2.0
multidict==6.0.5
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nbsphinx==0.9.4
ndx-grayscalevolume==0.0.2
ndx-icephys-meta==0.1.0
ndx-spectrum==0.2.2
nengo==4.0.0
nengo-gui==0.5.0
nengo-loihi==1.1.0
nengo-ocl==3.0.0
nengo-spa==2.0.0
nengo-sphinx-theme==20.9
nir==1.0.4
nirtorch==1.0
nlb-tools==0.0.3
notebook==7.2.0
notebook-shim==0.2.4
numba==0.59.1
numcodecs==0.12.1
numpydoc==1.7.0
nwbinspector==0.4.35
nwbwidgets==0.11.3
optuna==4.0.0
overrides==7.7.0
pandocfilters==1.5.1
pillow==10.3.0
plotly==5.13.1
pluggy==1.5.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
py2vega==0.6.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic-core==2.18.2
pynwb==2.7.0
pyopencl==2024.2.5
pyout==0.7.3
pyparsing==3.1.2
pytest==8.2.1
pytest-allclose==1.0.0
pytest-plt==1.1.1
pytest-rng==1.0.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pythreejs==2.4.2
pytools==2024.1.3
pytz==2024.1
pyyaml==6.0.1
qtconsole==5.5.2
qtpy==2.4.1
referencing==0.35.1
requests==2.32.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987==1.3.8
rpds-py==0.18.1
ruamel-yaml==0.18.6
s3fs==2024.5.0
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.13.1
semantic-version==2.10.0
send2trash==1.8.3
sniffio==1.3.1
snntorch==0.9.1
snowballstemmer==2.2.0
soupsieve==2.5
sparse==0.15.4
sphinx==7.3.7
sphinx-notfound-page==1.0.2
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlalchemy==2.0.34
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.3.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.5.22
tinycss2==1.3.0
tomli==2.0.1
torch==2.4.1
tqdm==4.66.4
traittypes==0.2.1
trimesh==4.4.0
types-python-dateutil==2.9.0.20240316
typing-extensions==4.12.0
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.18
webcolors==1.13
websocket-client==1.8.0
widgetsnbextension==4.0.10
wrapt==1.16.0
yarl==1.9.4
zarr==2.18.1
zarr-checksum==0.4.0
zipp==3.18.2